requirements.txt
biopython
pandas
torch
tqdm
numpy
matplotlib
sklearn